리팩토링: 함수를 실행하는 코드를 분리함.
